New and enhanced scripts for debugging
Enhances the functionality of the existing scripts for showing all
methods implemented by a class or trait, showing all messages sent by
the methods of a class or trait, showing all method conflicts in the set
of traits used by a class or trait and showing all messages sent to self
by the methods of a class for which the class has not implementing
method.

Added the script ShowAllSenders which shows all the methods of a class
or trait that send a particular message.